Change int size to size_t size throughout.  Fix int overflow for large messages.

+++ generic/com_mpi.c
@@ -209,7 +209,7 @@ typedef struct gather_t {
 /* structure to keep track of details of a declared gather */
 typedef struct gmem_t {
   char *mem;            /* source (destination) address for send (receive) */
-  int size;             /* size of sent field */
+  size_t size;          /* size of sent field */
   int stride;           /* stride of source/destination field */
   int num;              /* number of sites in sitelist */
   int *sitelist;        /* sites gathered to/from */
@@ -975,7 +975,7 @@ broadcast_dcomplex(double_complex *cpt)
 **  Broadcast bytes from node 0 to all others
 */
 void
-broadcast_bytes(char *buf, int size)
+broadcast_bytes(char *buf, size_t size)
 {
   MPI_Bcast( buf, size, MPI_BYTE, 0, MPI_COMM_THISJOB );
 }
@@ -1015,7 +1015,7 @@ receive_integer(int fromnode, int *address)
 **  send_field is to be called only by the node doing the sending
 */
 void
-send_field(char *buf, int size, int tonode)
+send_field(char *buf, size_t size, int tonode)
 {
   MPI_Send( buf, size, MPI_BYTE, tonode, SEND_FIELD_ID, MPI_COMM_THISJOB );
 }
@@ -1024,7 +1024,7 @@ send_field(char *buf, int size, int tonode)
 **  get_field is to be called only by the node to which the field was sent
 */
 void
-get_field(char *buf, int size, int fromnode)
+get_field(char *buf, size_t size, int fromnode)
 {
   MPI_Status status;
   MPI_Recv( buf, size, MPI_BYTE, fromnode, SEND_FIELD_ID, MPI_COMM_THISJOB,
@@ -1731,7 +1731,7 @@ msg_tag *
 declare_strided_gather(
   void *field,	        /* source buffer aligned to desired field */
   int stride,           /* bytes between fields in source buffer */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int subl,		/* subl of sites whose neighbors we gather.
@@ -2115,7 +2115,7 @@ cleanup_gather(msg_tag *mtag)
 msg_tag *
 declare_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2132,7 +2132,7 @@ declare_gather_site(
 msg_tag *
 start_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2157,7 +2157,7 @@ start_gather_site(
 void
 restart_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2203,7 +2203,7 @@ restart_gather_site(
 msg_tag *
 declare_gather_field(
   void * field,		/* which field? Pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2219,7 +2219,7 @@ declare_gather_field(
 msg_tag *
 start_gather_field(
   void * field,		/* which field? Pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2243,7 +2243,7 @@ start_gather_field(
 void
 restart_gather_field(
   void *field,		/* which field? Pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2450,7 +2450,7 @@ declare_accumulate_strided_gather(
   msg_tag **mmtag,      /* tag to accumulate gather into */
   void *field,	        /* which field? Some member of structure "site" */
   int stride,           /* bytes between fields in source buffer */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2475,7 +2475,7 @@ void
 declare_accumulate_gather_site(
   msg_tag **mmtag,
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2493,7 +2493,7 @@ void
 declare_accumulate_gather_field(
   msg_tag **mmtag,
   void * field,		/* which field? Pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -2549,7 +2549,7 @@ msg_tag *
 start_general_strided_gather(
   char *field,	        /* source buffer aligned to desired field */
   int stride,           /* bytes between fields in source buffer */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int *displacement,	/* displacement to gather from. four components */
   int parity,		/* parity of sites to which we gather.
 			   one of EVEN, ODD or EVENANDODD. */
@@ -2769,7 +2769,7 @@ msg_tag *
 start_general_strided_gather(
   char *field,	        /* source buffer aligned to desired field */
   int stride,           /* bytes between fields in source buffer */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int *displacement,	/* displacement to gather from. four components */
   int subl,		/* subl of sites whose neighbors we gather.
 			   It is EVENANDODD, if all sublattices are done. */
@@ -3093,7 +3093,7 @@ start_general_strided_gather(
 msg_tag *
 start_general_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int *displacement,	/* displacement to gather from. four components */
   int parity,		/* parity of sites to which we gather.
 			   one of EVEN, ODD or EVENANDODD. */
@@ -3106,7 +3106,7 @@ start_general_gather_site(
 msg_tag *
 start_general_gather_field(
   void * field,	        /* which field? Pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int *displacement,	/* displacement to gather from. four components */
   int parity,		/* parity of sites to which we gather.
 			   one of EVEN, ODD or EVENANDODD. */
