+++ include/comdefs.h
@@ -77,11 +77,11 @@ void broadcast_float(Real *fpt);
 void broadcast_double(double *dpt);
 void broadcast_complex(complex *cpt);
 void broadcast_dcomplex(double_complex *cpt);
-void broadcast_bytes(char *buf, int size);
+void broadcast_bytes(char *buf, size_t size);
 void send_integer(int tonode, int *address);
-void send_field(char *buf, int size, int tonode);
+void send_field(char *buf, size_t size, int tonode);
 void receive_integer(int fromnode, int *address);
-void get_field(char *buf, int size, int fromnode);
+void get_field(char *buf, size_t size, int fromnode);
 
 double dclock_cpu(void);
 double dclock(void);
@@ -101,7 +101,7 @@ int make_gather(
 
 msg_tag * declare_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -115,7 +115,7 @@ void cleanup_gather(msg_tag *mbuf);
 
 msg_tag * start_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -124,7 +124,7 @@ msg_tag * start_gather_site(
 
 void restart_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -134,7 +134,7 @@ void restart_gather_site(
 
 msg_tag * declare_gather_field(
   void * field,		/* which field? pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -144,7 +144,7 @@ msg_tag * declare_gather_field(
 msg_tag * declare_strided_gather(
   void *field,	        /* source buffer aligned to desired field */
   int stride,           /* bytes between fields in source buffer */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int subl,		/* subl of sites whose neighbors we gather.
@@ -153,7 +153,7 @@ msg_tag * declare_strided_gather(
 
 msg_tag * start_gather_field(
   void * field,		/* which field? pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -162,7 +162,7 @@ msg_tag * start_gather_field(
 
 void restart_gather_field(
   void * field,		/* which field? pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -177,7 +177,7 @@ void accumulate_gather(
 void declare_accumulate_gather_site(
   msg_tag **mmtag,      /* msg_tag to accumulate into */
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -187,7 +187,7 @@ void declare_accumulate_gather_site(
 void declare_accumulate_gather_field(
   msg_tag **mmtag,      /* msg_tag to accumulate into */
   void * field,		/* which field? pointer returned by malloc() */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int index,		/* direction to gather from. eg XUP - index into
 			   neighbor tables */
   int parity,		/* parity of sites whose neighbors we gather.
@@ -197,7 +197,7 @@ void declare_accumulate_gather_field(
 msg_tag * start_general_strided_gather(
   char *field,	        /* source buffer aligned to desired field */
   int stride,           /* bytes between fields in source buffer */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int *displacement,	/* displacement to gather from. four components */
   int subl,		/* subl of sites whose neighbors we gather.
 			   It is EVENANDODD, if all sublattices are done. */
@@ -205,7 +205,7 @@ msg_tag * start_general_strided_gather(
 
 msg_tag * start_general_gather_site(
   field_offset field,	/* which field? Some member of structure "site" */
-  int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+  size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
   int *displacement,	/* displacement to gather from. four components */
   int parity,		/* parity of sites to which we gather.
 			   one of EVEN, ODD or EVENANDODD. */
@@ -214,7 +214,7 @@ msg_tag * start_general_gather_site(
 msg_tag * start_general_gather_field(
 /* arguments */
  void * field,	        /* which field? Pointer returned by malloc() */
- int size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
+ size_t size,		/* size in bytes of the field (eg sizeof(su3_vector))*/
  int *displacement,	/* displacement to gather from. four components */
  int parity,		/* parity of sites to which we gather.
 			   one of EVEN, ODD or EVENANDODD. */
